requirements.txt
django-ckeditor-5>=0.2.0
Pillow>=10.0.0
django-unfold>=0.8.0

